requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pyyaml==6.0.1

# API
fastapi==0.110.0
uvicorn==0.27.1

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
#selenium==4.15.0
#scrapy==2.11.0
#playwright==1.40.0

# PDF processing
pypdf2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.8

# Text processing
spacy==3.7.2
nltk==3.8.1
langdetect==1.0.9
unidecode==1.3.7

# Embeddings & Vector stores
openai==1.6.1
sentence-transformers==3.3.1
# Keep transformers/hub compatible with sentence-transformers on Py3.12
transformers==4.48.3
huggingface-hub==0.28.1
google-generativeai==0.8.3
#pinecone-client==3.0.0
qdrant-client==1.7.0

# Search
elasticsearch==8.11.0
rank-bm25==0.2.2

# RAG frameworks
langchain==0.1.0
langchain-community==0.0.10
llama-index==0.9.30

# Zoho integrations
#zoho-crm-python==1.0.0
#requests-oauthlib==1.3.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Data processing
pandas==2.1.4
numpy==1.26.2